refactor: extract team selection
